IntegerField is too small too

+++ app/models.py
@@ -5,4 +5,4 @@ from django.db import models
 class MailSender(models.Model):
     rec_email = models.CharField(max_length=150)
     mailtext = models.TextField()
-    time_to_send = models.IntegerField()
+    time_to_send = models.BigIntegerField()
